teams-bot/requirements.txt
botbuilder-integration-aiohttp>=4.17.0
botbuilder-dialogs>=4.16.2
databricks-sdk[openai]==0.58.0
httpx
databricks-ai-bridge